Arreglada la base y se cambiaron procedimientos

+++ Scripts de SQL/grupoxestudiante_actualizar_nota_acumulada.sql
@@ -1,6 +1,6 @@
 use BDD1;
 go
-CREATE PROCEDURE grupoxestudiante_actualizar_nota_acumulada @id int, @nota int
+CREATE PROCEDURE grupoxestudiante_actualizar_nota_acumulada @id int, @nota decimal(7, 4)
 AS
 BEGIN
 	UPDATE dbo.GrupoxEstudiante as ge
